Extract pod spec creation into the core pkg
