+++ internal/resources/otelcollector/agent/resources.go
@@ -2,21 +2,18 @@ package agent
 
 import (
 	collectorconfig "github.com/kyma-project/telemetry-manager/internal/otelcollector/config"
+	"github.com/kyma-project/telemetry-manager/internal/resources/otelcollector/core"
 	"gopkg.in/yaml.v3"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/utils/pointer"
 )
 
 const (
 	configMapKey            = "relay.conf"
 	configHashAnnotationKey = "checksum/config"
-	collectorUser           = 10001
-	collectorContainerName  = "collector"
 )
 
 var (
@@ -66,7 +63,7 @@ func MakeConfigMap(config Config, collectorConfig collectorconfig.Config) *corev
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      config.BaseName,
 			Namespace: config.Namespace,
-			Labels:    makeDefaultLabels(config),
+			Labels:    core.MakeDefaultLabels(config.BaseName),
 		},
 		Data: map[string]string{
 			configMapKey: confYAML,
@@ -75,8 +72,10 @@ func MakeConfigMap(config Config, collectorConfig collectorconfig.Config) *corev
 }
 
 func MakeDaemonSet(config Config, configHash string) *appsv1.DaemonSet {
-	labels := makeDefaultLabels(config)
+	labels := core.MakeDefaultLabels(config.BaseName)
 	annotations := makePodAnnotations(configHash)
+	resources := corev1.ResourceRequirements{}
+	podSpec := core.MakePodSpec(config.BaseName, config.DaemonSet.Image, config.DaemonSet.PriorityClassName, resources)
 
 	return &appsv1.DaemonSet{
 		ObjectMeta: metav1.ObjectMeta{
@@ -93,83 +92,7 @@ func MakeDaemonSet(config Config, configHash string) *appsv1.DaemonSet {
 					Labels:      labels,
 					Annotations: annotations,
 				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Name:  collectorContainerName,
-							Image: config.DaemonSet.Image,
-							Args:  []string{"--config=/conf/" + configMapKey},
-							EnvFrom: []corev1.EnvFromSource{
-								{
-									SecretRef: &corev1.SecretEnvSource{
-										LocalObjectReference: corev1.LocalObjectReference{
-											Name: config.BaseName,
-										},
-										Optional: pointer.Bool(true),
-									},
-								},
-							},
-							Env: []corev1.EnvVar{
-								{
-									Name: "MY_POD_IP",
-									ValueFrom: &corev1.EnvVarSource{
-										FieldRef: &corev1.ObjectFieldSelector{
-											FieldPath:  "status.podIP",
-											APIVersion: "v1",
-										},
-									},
-								},
-								{
-									Name: "MY_NODE_NAME",
-									ValueFrom: &corev1.EnvVarSource{
-										FieldRef: &corev1.ObjectFieldSelector{
-											FieldPath: "spec.nodeName",
-										},
-									},
-								},
-							},
-							SecurityContext: &corev1.SecurityContext{
-								Privileged:               pointer.Bool(false),
-								RunAsUser:                pointer.Int64(collectorUser),
-								RunAsNonRoot:             pointer.Bool(true),
-								ReadOnlyRootFilesystem:   pointer.Bool(true),
-								AllowPrivilegeEscalation: pointer.Bool(false),
-								SeccompProfile: &corev1.SeccompProfile{
-									Type: corev1.SeccompProfileTypeRuntimeDefault,
-								},
-								Capabilities: &corev1.Capabilities{
-									Drop: []corev1.Capability{"ALL"},
-								},
-							},
-							VolumeMounts: []corev1.VolumeMount{{Name: "config", MountPath: "/conf"}},
-							LivenessProbe: &corev1.Probe{
-								ProbeHandler: corev1.ProbeHandler{
-									HTTPGet: &corev1.HTTPGetAction{Path: "/", Port: intstr.IntOrString{IntVal: 13133}},
-								},
-							},
-							ReadinessProbe: &corev1.Probe{
-								ProbeHandler: corev1.ProbeHandler{
-									HTTPGet: &corev1.HTTPGetAction{Path: "/", Port: intstr.IntOrString{IntVal: 13133}},
-								},
-							},
-						},
-					},
-					ServiceAccountName: config.BaseName,
-					PriorityClassName:  config.DaemonSet.PriorityClassName,
-					Volumes: []corev1.Volume{
-						{
-							Name: "config",
-							VolumeSource: corev1.VolumeSource{
-								ConfigMap: &corev1.ConfigMapVolumeSource{
-									LocalObjectReference: corev1.LocalObjectReference{
-										Name: config.BaseName,
-									},
-									Items: []corev1.KeyToPath{{Key: configMapKey, Path: configMapKey}},
-								},
-							},
-						},
-					},
-				},
+				Spec: podSpec,
 			},
 		},
 	}
@@ -183,10 +106,4 @@ func makePodAnnotations(configHash string) map[string]string {
 		annotations[k] = v
 	}
 	return annotations
-}
-
-func makeDefaultLabels(config Config) map[string]string {
-	return map[string]string{
-		"app.kubernetes.io/name": config.BaseName,
-	}
 }
